--- mirror/filesystem_test.go
@@ -1,1 +1,0 @@
-package mirror
